requirements.txt
tensorflow==2.0.0
keras==2.3.1
Django==3.0.6
djangorestframework==3.11.0
drf-yasg==1.17.1
numpy==1.18.4
pandas==1.0.3